Colocate ScoresExportContext with scores export module.
Per-analytic service bundles belong next to the analytic, not in the shared export query context.
