+++ packages/api/api/analytics/export_context.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from collections.abc import Callable, Mapping
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
 from api.analytics.export_dependency_walk import (
     DependencyWalkResult,
@@ -26,23 +26,7 @@ from api.analytics.exports.jsonpath import parse_jsonpath, resolve_jsonpath
 from api.analytics.options import TurnAnalyticsOptions
 from api.models.game import TurnInfo
 
-if TYPE_CHECKING:
-    from api.analytics.military_score_inference.hull_catalog_mask import ResolvedHullCatalogMask
-    from api.analytics.military_score_inference.inference_scheduler import InferenceRowScheduler
-    from api.services.inference_row_persistence_service import InferenceRowPersistenceService
-
 INLINE_ENSURE_MAX_MISSING_STEPS = 5
-
-
-@dataclass(frozen=True)
-class ScoresExportContext:
-    """Inference services used by scores export ensure and materialization."""
-
-    persistence: InferenceRowPersistenceService | None = None
-    scheduler: InferenceRowScheduler | None = None
-    resolve_hull_catalog_mask: Callable[[TurnInfo, int], ResolvedHullCatalogMask | None] | None = (
-        None
-    )
 
 
 @dataclass(frozen=True)
