+++ packages/api/tests/scores_exports_helpers.py
@@ -7,7 +7,8 @@ from dataclasses import replace
 from pathlib import Path
 
 import pytest
-from api.analytics.export_context import ScoresExportContext, make_analytic_query_context
+from api.analytics.export_context import make_analytic_query_context
+from api.analytics.scores.export_services import ScoresExportContext
 from api.analytics.military_score_inference.inference_scheduler import InferenceRowScheduler
 from api.analytics.military_score_inference.inference_stream_rows import schedule_inference_row
 from api.analytics.military_score_inference.inference_stream_scope import InferenceStreamScope
